+++ roqoqo/src/operations/define_operations.rs
@@ -31,6 +31,8 @@
 //! (5) InputSymbolic, where the user can define a floating point type value to replace a certain symbolic parameter.
 //!
 
+use std::collections::HashSet;
+
 use crate::operations::{Define, InvolveQubits, InvolvedQubits, Operate, RoqoqoError, Substitute};
 
 /// DefinitionFloat is the Definition for a floating point type register.
@@ -60,9 +62,11 @@ const TAGS_DefinitionFloat: &[&str; 3] = &["Operation", "Definition", "Definitio
 
 // Implementing the InvolveQubits trait for DefinitionFloat.
 impl InvolveQubits for DefinitionFloat {
-    /// Lists all involved Qubits (here, none).
     fn involved_qubits(&self) -> InvolvedQubits {
         InvolvedQubits::None
+    }
+    fn involved_classical(&self) -> super::InvolvedClassical {
+        super::InvolvedClassical::All(self.name.clone())
     }
 }
 
@@ -99,6 +103,10 @@ impl InvolveQubits for DefinitionComplex {
     fn involved_qubits(&self) -> InvolvedQubits {
         InvolvedQubits::None
     }
+
+    fn involved_classical(&self) -> super::InvolvedClassical {
+        super::InvolvedClassical::All(self.name.clone())
+    }
 }
 
 /// DefinitionUsize is the Definition for an Integer type register.
@@ -131,6 +139,10 @@ impl InvolveQubits for DefinitionUsize {
     /// Lists all involved Qubits (here, none).
     fn involved_qubits(&self) -> InvolvedQubits {
         InvolvedQubits::None
+    }
+
+    fn involved_classical(&self) -> super::InvolvedClassical {
+        super::InvolvedClassical::All(self.name.clone())
     }
 }
 
@@ -165,6 +177,10 @@ impl InvolveQubits for DefinitionBit {
     fn involved_qubits(&self) -> InvolvedQubits {
         InvolvedQubits::None
     }
+
+    fn involved_classical(&self) -> super::InvolvedClassical {
+        super::InvolvedClassical::All(self.name.clone())
+    }
 }
 
 /// InputSymbolic is the Definition for a floating point type parameter which will replace a certain symbolic parameter.
@@ -194,5 +210,50 @@ impl InvolveQubits for InputSymbolic {
     /// Lists all involved Qubits (here, none).
     fn involved_qubits(&self) -> InvolvedQubits {
         InvolvedQubits::None
+    }
+
+    fn involved_classical(&self) -> super::InvolvedClassical {
+        let mut a: HashSet<(String, usize)> = HashSet::new();
+        a.insert((self.name.clone(), 0));
+        super::InvolvedClassical::Set(a)
+    }
+}
+
+/// InputBit sets a certain bit in an existing BitRegister of the circuit.
+///
+#[derive(
+    Debug,
+    Clone,
+    PartialEq,
+    Eq,
+    roqoqo_derive::Operate,
+    roqoqo_derive::Substitute,
+    roqoqo_derive::Define,
+)]
+#[cfg_attr(feature = "serialize", derive(serde::Serialize, serde::Deserialize))]
+// #[cfg_attr(feature = "json_schema", derive(schemars::JsonSchema))]
+pub struct InputBit {
+    /// The name of the register that where the bit is set.
+    name: String,
+    /// The index in the register that is set.
+    index: usize,
+    /// The value the bit is set to
+    value: bool,
+}
+
+#[allow(non_upper_case_globals)]
+const TAGS_InputBit: &[&str; 3] = &["Operation", "Definition", "InputBit"];
+
+// Implementing the InvolveQubits trait for InputSymbolic.
+impl InvolveQubits for InputBit {
+    /// Lists all involved Qubits (here, none).
+    fn involved_qubits(&self) -> InvolvedQubits {
+        InvolvedQubits::None
+    }
+
+    fn involved_classical(&self) -> super::InvolvedClassical {
+        let mut a: HashSet<(String, usize)> = HashSet::new();
+        a.insert((self.name.clone(), self.index));
+        super::InvolvedClassical::Set(a)
     }
 }